Refactored crappychat plugins to make room for dataset mentions.
