+++ src/Components/CrappyChat/EntryBox.js
@@ -1,6 +1,5 @@
 import React from "react";
 import Editor from "draft-js-plugins-editor";
-import { fromJS } from "immutable";
 import {
   EditorState,
   ContentState,
@@ -9,8 +8,8 @@ import {
 } from "draft-js";
 import base from "./Base";
 
-import pluginsFn from "./Plugins";
-import Entry from "./AtMentionEntry";
+import pluginsFn from "./Plugins/Plugins";
+import PluginComponents from './Plugins/PluginComponents';
 
 export default class EntryBox extends React.Component {
   constructor(props) {
@@ -20,7 +19,6 @@ export default class EntryBox extends React.Component {
 
     this.state = {
       editorState: EditorState.createEmpty(),
-      suggestions: fromJS([]),
       users: []
     };
   }
@@ -50,32 +48,7 @@ export default class EntryBox extends React.Component {
     this.resetState();
   }
 
-  onSearchChange({ value }) {
-    base
-      .fetch("users", {
-        context: this,
-        asArray: true
-      })
-      .then(users => {
-        this.setState({
-          suggestions: fromJS(
-            users
-              .filter(
-                user =>
-                  user.displayName.toLowerCase().indexOf(value.toLowerCase()) >
-                  -1
-              )
-              .map(user => ({
-                ...user,
-                name: user.displayName
-              }))
-          )
-        });
-      });
-  }
-
   render() {
-
     return (
       <form onSubmit={this.onSubmit.bind(this)}>
         <Editor
@@ -83,11 +56,7 @@ export default class EntryBox extends React.Component {
           onChange={this.onEditorChange.bind(this)}
           plugins={Object.values(this.plugins)}
         />
-        <this.plugins.mention.MentionSuggestions
-          onSearchChange={this.onSearchChange.bind(this)}
-          suggestions={this.state.suggestions}
-          entryComponent={Entry}
-        />
+        <PluginComponents userMentionsPlugin={this.plugins.userMentions} />
         <input type="submit" />
       </form>
     );
