+++ src/Components/CrappyChat/Plugins/UserMention/UserMentionEntry.js
